Revert "Split serialized state types into their own file"
This reverts commit 4c9c5b99d32949b26e87be9eabd818c74df8b81f.
