+++ js/reducers/playlist.ts
@@ -28,7 +28,9 @@ import { shuffle, moveSelected, objectMap, objectFilter } from "../utils";
 
 export interface PlaylistState {
   trackOrder: number[];
-  tracks: { [id: string]: PlaylistTrack };
+  // https://github.com/Microsoft/TypeScript/pull/12253#issuecomment-263132208
+  // TODO: Using numbers for keys is kinda annoying. Consider retyping as string
+  tracks: { [id: number]: PlaylistTrack };
   lastSelectedIndex: number | null;
   currentTrack: number | null;
 }
